requirements.txt
streamlit
librosa
numpy
matplotlib
scikit-learn
joblib
torch
torchvision
pillow